Minilogo stuff on Shallow Grave, along with chang-ing how it modifies damage. Animation stuff for missing. Minilogos on all out-of-turn attacks.

+++ Utopia/Classes/Battle/Skills/SkillPoisonFire.h
@@ -10,8 +10,8 @@
 
 @interface SkillPoisonFire : SkillControllerSpecialOrb
 {
-  int _poisonStacks;
-  int _quickAttackStacks;
+  NSInteger _poisonStacks;
+  NSInteger _quickAttackStacks;
   int _initialDamage;
   int _poisonDamage;
   int _poisonPercent;
